feat: implementation of detection vulnerabilities
Description: this is an implementation trough Selenium of a testing mechanism that detect if a site is vulnerabilities of the ClickJacking described in Owasp, this class permit to create an environment before the detection of vulnerability trough the method creationEvnviroment(), in which are generated the html file that implement the clickjacking attack, and then we can select the clickJaking attack selected by the user trough the method selecteTack(), and test trough the method executioneTest()  that attack the site passed by the, and after the detection if the site is vulnerable or not, we can clean the environment with the delete of the html files trough method cleanEnvironment(). An important thing to notice is the Thread sleep line 119 it useful because we can wait time that the html file dynamically created, will be deployed on the Apache tomcat  server.