refactor: Change path File, add class ="malicious"
description:
Change Row 17:
I have changed the path in which the will be written, in particular the
Html file will be written in a dedicated folder "html_generated" to
better organization of the project.

Change Row 46: I have added class="malicious" that is assingned to the
iframe, because if we have a nested chain of iframe so an iframe that
call another another html with an iframe and so on, with this
assignement and the function SwitchTo of Selenium we can switch in the
neasted chain of iframe.