refactor: Catch Exception, increse of thread sleep, assignement return
description:
Line 120 i have incremented the thread sleep to be better sure that the
server has the time to syncronize on it the html dynamically generated.
Line 146 i have added the automatically exit  from the chrome browser
after the test of site.
Line 164 i have added the catching of another exception because if the
site block the clickjacking attack DisablingJavascript on explorer
browser, explorer generate an exception that is JavascriptException, and
so i have to catch it.
Line 168 i have added the automatically exit  from the Internet Explorer
browser after the test of site.
Line 170-172 i have added the assignment of result to array results[]
one time, before this operation was duplicated at line 148 and 173