refactoring: Decoupling assignment of path
description: This class permit to decouple the assignment of path file
in the project, indeed previously when we change the path of the
execution of the project, for instance if we change the machine of the
execution, we have must change the path in the classes
"RedefiningLocation", "OnBeforeUnloadEvent",
"DisablingJavascriptChrome", "DisablingJavascriptIE", "FilterXssIE",
"FilterXssChrome", "Fictitious1", "Fictitious2", "TestingClickJacking".
Trough this class FileCustom if we change the path of the execution of
the project must change only the path in this class in constructor with
path on where is the project.