improvement: improve wait of synchronization of server, read csv path
description:
Form line 130-138:
i have improved the mechanism of wait of the deployment of the html file
on server, indeed in the previously code at line 123 there was a wait
statically of 10 seconds, to improve this mechanism in more smart way i
have exploit the class TestPageOk() described in the commit:
"8505dadf778fe6eac58334bd0edee21b4afbdfdd", indeed now as we can see in
the line 131 in the file modified, we instantiate the class TestPage
passing as parameters, the url on which is loaded the html page
dynamically created, and the number of tentative, that we want to make
to do, to see if the process of loading of the html page dynamically
created give is all ok. So in this way when we execute at line 133 the
method testPageExist() we make "n" tentative and each tentative wait
seconds before to make another, and in each tentative the code try to
load the page at the url passed, and if one of the n response is 202, is
all ok, if after the n tentative there is a response different from 202
there is the catching of exception and so we have the print of the HTTP
response code to understand what is the issue. Another important
improvement of this mechanism is that before, the wait mechanism with
Thread.sleep() for the synchronization with the deployment of the file
pn the server, is applied directly in this class now all the mechanism
is made in the TestPageOk so this increase the decoupling of
responsibility increasing the maintainability of the code, and the reuse
of the mechanism implemented in the class TestPageOKthis in other
situation.

At line 194: Through the class FileCustom described in the commit
"c3e8949ef922487bf9a474d05ca9a5f389fffcbe" we don't have to anymore
change the path in this line if we change the execution path of the
project.