Implementation of X-Frame-Option
description: implementation of technique of defense from Clickjacking
X-Frame-Option.
In the file app.js we use Node.js to create and manage a server.
lines 45-57 instantiate the server, lines 30-43 when receive a HTTP
request from frame.html(the site attacked) it set header
'X-Frame-Options: DENY' if receive 1 from command line, or if receive 2
set the header with 'X-Frame-Options: SAMERORIGIN'.

The file frame.html is the site attacked, that is referred in the iframe
of the attacker site.

The file index.html is the site that attack, that contain in the iframe,
the reference to the file frame.html, line 8.