Improvement: improved the mechanism of detecting if a site is vulnerable
description: improvement of the mechanism of checking whether a site is vulnerable or not from Clickjacking attack. The checking now is made trough the method "takingHref(..)", that is called on the original site without attack, and on the site attacked that has attacked the original site, and this method return the links "<link... href.. stylesheet>" of the two sites, so that the links that refer to external CSS document, after is called method "takingStyle(.....)" that is called on the original site without attack, and on the site attacked that has attacked the original site, this method return if the two sites contain the tag style, that refer to the fact that if the site has loaded CSS elements or not in it, after is called method comparisonResult(.....), that compare if the original site without attack and the site that has attacked the original contain or not, both the tag "style" and the same links "<link... href.. stylesheet>", so this checking mechanism is based on the fact that a site in an iframe is loaded correctly so is vulnerable, if all the CSS elements and references to CSS document contained in it, are correctly loaded.