refactoring: avoid redundant code, handle some exception
description: I have deleted the redundant code, indeed before there was a code clone as we can see in the previous code lines 233-240, 249-253 these codes made the same thing, that was the adding to an arraylist of String  the href of web element <link rel="stylesheet"...>, and now i have implemented this functionality only one time lines 252-279. In addition in these 252-279 lines i have resolved the problem that sometimes occur about NullPointerException when for example an element link in an html file doesn't have the attribute rel or href in these cases there is a reference to a null object, so with catch(NullPointerException), we can capture this problem and handle it in a way that only if the web element contain rel="stylesheet" and href"...." attribute we can add its href to the arraylist of String returned by the method takingHref(....).